Add redis store for tracking chosen players.
Also lays the foundation for making the page dynamic.
The routes are in place now we just need the js to poll /game/:gameId/info.
The foundation for number of players is now laid as well we just need to do a PUT to /game/:gameId with the new count.